+++ src/object/tcb.c
@@ -767,8 +767,8 @@ decodeDomainInvocation(word_t label, unsigned int length, extra_caps_t extraCaps
     } else {
         domain = getSyscallArg(0, buffer);
         if (domain >= CONFIG_NUM_DOMAINS) {
-            userError("Domain Configure: invalid domain (%d >= %d).",
-                      (int)domain, CONFIG_NUM_DOMAINS);
+            userError("Domain Configure: invalid domain (%u >= %u).",
+                      domain, CONFIG_NUM_DOMAINS);
             current_syscall_error.type = seL4_InvalidArgument;
             current_syscall_error.invalidArgumentNumber = 0;
             return EXCEPTION_SYSCALL_ERROR;
